specops/pyproject.toml
[project]
name = "specops"
dynamic = ["version"]
description = "Multi-agent team admin and control plane"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Hung Tran", email = "hungtd9@gmail.com"}
]
keywords = ["ai", "agent", "multi-agent", "team", "automation", "admin", "control-plane"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: FastAPI",
]
dependencies = [
    "specops_lib[all-channels]",
    "fastapi>=0.115.0",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.30.0",
    "PyJWT>=2.8.0",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "docker>=7.0.0",
]

[project.optional-dependencies]
s3 = ["specops_lib[s3]"]

[project.scripts]
specops = "specops.cli:app"

[project.urls]
Homepage = "https://github.com/taylorelley/specops"
Documentation = "https://github.com/taylorelley/specops#readme"
Repository = "https://github.com/taylorelley/specops"
Issues = "https://github.com/taylorelley/specops/issues"

[tool.uv.sources]
specops_lib = { workspace = true }

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0.dev0"

[tool.hatch.build.targets.wheel]
packages = ["specops"]

[tool.hatch.build]
include = [
    "specops/**/*.py",
    "specops/templates/**/*.md",
]
